Add tinydocker.py
